remove the url utility
