+++ tests/FunctionalTests/VerifyEmailHelperFunctionalTest.php
@@ -15,7 +15,6 @@ use Symfony\Component\HttpKernel\UriSigner;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use SymfonyCasts\Bundle\VerifyEmail\Generator\VerifyEmailTokenGenerator;
 use SymfonyCasts\Bundle\VerifyEmail\Util\VerifyEmailQueryUtility;
-use SymfonyCasts\Bundle\VerifyEmail\Util\VerifyEmailUrlUtility;
 use SymfonyCasts\Bundle\VerifyEmail\VerifyEmailHelper;
 use SymfonyCasts\Bundle\VerifyEmail\VerifyEmailHelperInterface;
 
@@ -108,7 +107,7 @@ class VerifyEmailHelperFunctionalTest extends TestCase
         return new VerifyEmailHelper(
             $this->mockRouter,
             new UriSigner('foo', 'signature'),
-            new VerifyEmailQueryUtility(new VerifyEmailUrlUtility()),
+            new VerifyEmailQueryUtility(),
             new VerifyEmailTokenGenerator('foo'),
             3600
         );
