+++ tests/UnitTests/Util/VerifyEmailQueryTest.php
@@ -9,24 +9,12 @@
 
 namespace SymfonyCasts\Bundle\VerifyEmail\Tests\UnitTests\Util;
 
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use SymfonyCasts\Bundle\VerifyEmail\Model\VerifyEmailUrlComponents;
 use SymfonyCasts\Bundle\VerifyEmail\Util\VerifyEmailQueryUtility;
-use SymfonyCasts\Bundle\VerifyEmail\Util\VerifyEmailUrlUtility;
 
 class VerifyEmailQueryTest extends TestCase
 {
-    /**
-     * @var MockObject|VerifyEmailUrlUtility
-     */
-    private $mockUrlUtility;
-
-    protected function setUp(): void
-    {
-        $this->mockUrlUtility = $this->createMock(VerifyEmailUrlUtility::class);
-    }
-
     public function testGetsExpiryTimeFromQueryString(): void
     {
         $uri = '/?a=x&expires=1234567890';
@@ -35,14 +23,14 @@ class VerifyEmailQueryTest extends TestCase
         $components->setPath('/');
         $components->setQuery('a=x&expires=1234567890');
 
-        $this->mockUrlUtility
-            ->expects($this->once())
-            ->method('parseUrl')
-            ->with($uri)
-            ->willReturn($components)
-        ;
+//        $this->mockUrlUtility
+//            ->expects($this->once())
+//            ->method('parseUrl')
+//            ->with($uri)
+//            ->willReturn($components)
+//        ;
 
-        $queryUtility = new VerifyEmailQueryUtility($this->mockUrlUtility);
+        $queryUtility = new VerifyEmailQueryUtility();
         $result = $queryUtility->getExpiryTimestamp($uri);
 
         self::assertSame(
